pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tortoise-orm[asyncpg,psycopg]>=0.25",
    "ruff>=0.11.0",
    "python-dotenv>=1.1.0",
    "aiohttp>=3.11.18",
    "anyio>=4.9.0",
    "pydantic>=2.11.3",
]


[tool.ruff]
indent-width = 2
